requirements.txt
aniso8601==7.0.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cryptography==3.4.8
Flask==2.0.1
Flask-Cors==3.0.10
Flask-GraphQL==2.0.1
flask-marshmallow==0.14.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==1.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
midtransclient==1.2.0
promise==2.3
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.1.0
PyMySQL==1.0.2
python-dotenv==0.19.0
redis==3.5.3
requests==2.26.0
RouterOS-api==0.17.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
SQLAlchemy==1.4.23
urllib3==1.26.6
Werkzeug==2.0.1
